Fix gce_ubuntu setup scripts. Close #46

+++ gce_ubuntu/setup_vnc.sh
@@ -6,24 +6,38 @@ set -o pipefail
 set -o xtrace
 
 readonly VNC_PASSWORD="${VNC_PASSWORD:-password}"
+readonly VNC_DISPLAY=':1'
 
 exist_vncserver() {
-  local vnc_server_id="$1"
-  [[ "$(vncserver -list | grep "${vnc_server_id}" | wc -l)" -ne 0 ]]
+  local vnc_display="$1"
+  [[ "$(vncserver -list | grep "${vnc_display}" | wc -l)" -ne 0 ]]
 }
 
-sudo apt update
+# state validation
+if exist_vncserver "${VNC_DISPLAY}"; then
+  echo "VNC display "${VNC_DISPLAY}" already exists."
+  exit 0
+fi
 
 # X Window System, GNOME, VNC Server
+sudo apt update
 sudo apt install -y ubuntu-desktop \
     gnome-core \
     gnome-panel \
     metacity \
     tigervnc-standalone-server
 
-# VNC Setting
-mkdir -p ~/.vnc
-echo "${VNC_PASSWORD}" | vncpasswd -f > ~/.vnc/passwd
+# set password
+cat << EOT | vncpasswd
+${VNC_PASSWORD}
+${VNC_PASSWORD}
+EOT
+
+exit 1
+
+# initialize
+vncserver "${VNC_DISPLAY}"
+vncserver -kill "${VNC_DISPLAY}"
 
 cat << EOT > ~/.vnc/xstartup
 exec gnome-session &
@@ -35,6 +49,4 @@ gnome-terminal &
 EOT
 
 # start
-if ! exist_vncserver :1; then
-  vncserver :1
-fi
+vncserver "${VNC_DISPLAY}"
